+++ src/tools/DrawingTools.js
@@ -1,7 +1,6 @@
 import EventEmitter from 'tiny-emitter';
 import RubberbandRectTool from './rectangle/RubberbandRectTool';
 import RubberbandPolygonTool from './polygon/RubberbandPolygonTool';
-import { SVG_NAMESPACE } from '../util/SVG';
 
 /** The drawing tool 'registry' **/
 class DrawingToolRegistry extends EventEmitter {
@@ -68,51 +67,3 @@ class DrawingToolRegistry extends EventEmitter {
 }
 
 export default DrawingToolRegistry;
-
-/**
- * Common code for drawing resize handles
- */
-export const drawHandle = (x, y) => {
-  const containerGroup = document.createElementNS(SVG_NAMESPACE, 'g');
-  containerGroup.setAttribute('class', 'a9s-handle');
-  containerGroup.setAttribute('transform-origin', `${x}px ${y}px`);
-
-  const group = document.createElementNS(SVG_NAMESPACE, 'g');
-  group.setAttribute('transform-origin', `${x}px ${y}px`);
-
-  const drawCircle = r => {
-    const c = document.createElementNS(SVG_NAMESPACE, 'circle');
-    c.setAttribute('cx', x);
-    c.setAttribute('cy', y);
-    c.setAttribute('r', r);
-    return c;
-  }
-
-  const inner = drawCircle(6);
-  inner.setAttribute('class', 'a9s-handle-inner')
-
-  const outer = drawCircle(7);
-  outer.setAttribute('class', 'a9s-handle-outer')
-
-  group.appendChild(outer);
-  group.appendChild(inner);
-
-  containerGroup.appendChild(group);
-  return containerGroup;
-}
-
-/**
- * Common code for setting handle position
- */
-export const setHandleXY = (handle, x, y) => {
-  handle.setAttribute('transform-origin', `${x}px ${y}px`);	
-  handle.firstChild.setAttribute('transform-origin', `${x}px ${y}px`);	
-
-  const inner = handle.querySelector('.a9s-handle-inner');	
-  inner.setAttribute('cx', x);	
-  inner.setAttribute('cy', y);	
-
-  const outer = handle.querySelector('.a9s-handle-outer');	
-  outer.setAttribute('cx', x);	
-  outer.setAttribute('cy', y);
-}
