+++ src/tools/polygon/EditablePolygon.js
@@ -2,7 +2,7 @@ import EventEmitter from 'tiny-emitter';
 import { drawEmbeddedSVG, toSVGTarget } from '../../selectors/EmbeddedSVG';
 import { SVG_NAMESPACE } from '../../util/SVG';
 import { format } from '../../util/Formatting';
-import { drawHandle, setHandleXY } from '../DrawingTools';
+import { drawHandle, setHandleXY } from '../Tool';
 import Mask from './PolygonMask';
 
 const getPoints = shape => {
